+++ src/backend/worker_analytics_service.js
