--- src/pages/worker/worker_dashboard.css
@@ -1,280 +1,0 @@
-.worker_page {
-	display: flex;
-	min-height: 100vh;
-	background: #f8fafc;
-	font-family: 'DM Sans', sans-serif;
-}
-
-.worker_sidebar {
-	width: 220px;
-	background: #0f1f3d;
-	display: flex;
-	flex-direction: column;
-	flex-shrink: 0;
-	min-height: 100vh;
-	padding: 0;
-}
-
-.worker_sidebar_logo {
-	font-family: 'Barlow Condensed', sans-serif;
-	font-size: 22px;
-	font-weight: 700;
-	color: #fff;
-	letter-spacing: 0.04em;
-	padding: 20px 20px 4px;
-}
-
-.worker_sidebar_logo span {
-	color: #f59e0b;
-}
-
-.worker_sidebar_role {
-	font-size: 11px;
-	color: rgba(255, 255, 255, 0.4);
-	padding: 0 20px 16px;
-	border-bottom: 1px solid rgba(255, 255, 255, 0.08);
-}
-
-.worker_sidebar_nav {
-	flex: 1;
-	padding: 12px 0;
-}
-
-.worker_sidebar_item {
-	display: flex;
-	align-items: center;
-	width: 100%;
-	padding: 9px 20px;
-	font-size: 13px;
-	color: rgba(255, 255, 255, 0.6);
-	border-left: 2px solid transparent;
-	cursor: pointer;
-}
-
-.worker_sidebar_item.active {
-	background: rgba(245, 158, 11, 0.15);
-	color: #f59e0b;
-	border-left: 2px solid #f59e0b;
-}
-
-.worker_sidebar_badge {
-	margin-left: auto;
-	background: #f59e0b;
-	color: #0f1f3d;
-	font-size: 11px;
-	font-weight: 600;
-	padding: 2px 7px;
-	border-radius: 99px;
-}
-
-.worker_sidebar_bottom {
-	padding: 16px 20px;
-	border-top: 1px solid rgba(255, 255, 255, 0.08);
-	display: flex;
-	align-items: center;
-	gap: 10px;
-}
-
-.worker_sidebar_avatar {
-	width: 32px;
-	height: 32px;
-	border-radius: 50%;
-	background: #1a3260;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	font-size: 12px;
-	font-weight: 500;
-	color: #f59e0b;
-	flex-shrink: 0;
-}
-
-.worker_sidebar_user_name {
-	font-size: 12px;
-	color: #fff;
-	font-weight: 500;
-}
-
-.worker_sidebar_user_role {
-	font-size: 11px;
-	color: rgba(255, 255, 255, 0.4);
-}
-
-.worker_main {
-	flex: 1;
-	display: flex;
-	flex-direction: column;
-}
-
-.worker_top_bar {
-	display: flex;
-	align-items: center;
-	justify-content: space-between;
-	padding: 20px 32px;
-	background: #fff;
-	border-bottom: 1px solid #e2e8f0;
-}
-
-.worker_top_bar_title {
-	font-size: 20px;
-	font-weight: 500;
-	color: #0f1f3d;
-	margin: 0;
-}
-
-.worker_top_bar_sub {
-	font-size: 13px;
-	color: #64748b;
-	margin: 2px 0 0;
-}
-
-.worker_signout_btn {
-	font-size: 13px;
-	padding: 8px 16px;
-	background: transparent;
-	border: 1px solid #e2e8f0;
-	border-radius: 6px;
-	color: #64748b;
-	cursor: pointer;
-}
-
-.worker_signout_btn:hover {
-	background: #f8fafc;
-	color: #0f1f3d;
-}
-
-.worker_content {
-	padding: 24px 32px;
-}
-
-.worker_filters {
-	display: flex;
-	gap: 8px;
-	margin-bottom: 24px;
-	flex-wrap: wrap;
-}
-
-.worker_filter_btn {
-	padding: 6px 14px;
-	border-radius: 99px;
-	border: 1px solid #e2e8f0;
-	background: #fff;
-	font-size: 12px;
-	color: #64748b;
-	cursor: pointer;
-	text-transform: capitalize;
-}
-
-.worker_filter_btn.active {
-	background: #0f1f3d;
-	color: #f59e0b;
-	border-color: #0f1f3d;
-}
-
-.worker_requests_grid {
-	display: grid;
-	grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
-	gap: 16px;
-}
-
-.worker_card_wrapper {
-	cursor: pointer;
-}
-
-.worker_loading,
-.worker_empty {
-	color: #64748b;
-	font-size: 14px;
-	margin-top: 40px;
-	text-align: center;
-}
-
-.worker_content .request_card {
-	background: #1e293b;
-	border: 1px solid #334155;
-	border-radius: 18px;
-	padding: 22px;
-	box-shadow: 0 4px 16px rgba(0, 0, 0, 0.2);
-	display: flex;
-	flex-direction: column;
-	gap: 12px;
-}
-
-.worker_content .request_card_top {
-	display: flex;
-	align-items: center;
-	justify-content: space-between;
-}
-
-.worker_content .request_card_top h3 {
-	margin: 0;
-	font-size: 1.35rem;
-	color: #f8fafc;
-	text-transform: capitalize;
-}
-
-.worker_content .request_location {
-	color: rgba(255, 255, 255, 0.5);
-	font-size: 13px;
-}
-
-.worker_content .request_description {
-	color: rgba(255, 255, 255, 0.8);
-	font-size: 13px;
-}
-
-.worker_unclaimed_card {
-	display: flex;
-	flex-direction: column;
-}
-
-.worker_claim_btn {
-	width: 100%;
-	padding: 10px;
-	background: #f59e0b;
-	color: #0f1f3d;
-	border: none;
-	border-radius: 8px;
-	font-size: 13px;
-	font-weight: 500;
-	cursor: pointer;
-	margin-top: 4px;
-}
-
-.worker_claim_btn:hover:not(:disabled) {
-	background: #d97706;
-}
-
-.worker_claim_btn:disabled {
-	opacity: 0.5;
-	cursor: not-allowed;
-}
-
-.status_badge {
-	font-size: 12px;
-	padding: 4px 10px;
-	border-radius: 99px;
-	font-weight: 500;
-	text-transform: capitalize;
-}
-
-.status_badge.submitted {
-	background: #2563eb;
-	color: #fff;
-}
-.status_badge.assigned {
-	background: #fef9c3;
-	color: #854d0e;
-}
-.status_badge.in_progress {
-	background: #fef3c7;
-	color: #d97706;
-}
-.status_badge.resolved {
-	background: #dcfce7;
-	color: #166534;
-}
-.status_badge.pending {
-	background: #2563eb;
-	color: #fff;
-}
